Connected map view/slide up view to swipe view
If a user selects an item in the popular places table from the slide up view, the user is segued to the swipe view. The swipe view is populated by the popular places list. The first view of all views in the swipe controller is the view of the place that was selected. A lot of UI changes still need to be done.

For some reason the pictures won't load???